Move assets schemas to serialization module
